requirements.txt
python-docx==1.1.2
rapidfuzz==3.9.7
requests==2.32.3
selenium==4.24.0
webdriver-manager==4.0.2
openai==1.52.2
langchain-openai==0.2.8
langchain-core==0.3.15
langchain-community==0.3.8
langchain-text-splitters==0.3.2
faiss-cpu==1.9.0
python-dotenv==1.0.1
